route from main page to search page
