+++ client/components/Home.jsx
@@ -1,22 +1,27 @@
 import React, { Component } from "react";
+import { useHistory } from "react-router-dom";
 
-class Home extends Component {
+function Home(props){
 
-  handleClick() {
-    this.props.history.push('/search');
+  let history = useHistory();
+
+  const handleClick = async () => {
+    await props.querySearch();
+    history.push('/search');
   }
 
-  render() {
-    return(
-        <div className="home-div">
-          <div className="home-logo">movienite</div>
-            <div className="home-searchbar">
-              <input className="search-input" type="text"></input>
-              <button onClick={() => this.handleClick()} className="SearchButton">Search</button>
-            </div>
-        </div>
-    )
-  }
+  const { query, updateQuery, querySearch } = props;
+
+  return(
+      <div className="home-div">
+        <div className="home-logo">movienite</div>
+          <div className="home-searchbar">
+            <input onChange={updateQuery} className="search-input" type="text" value={query}></input>
+            <button onClick={handleClick} className="SearchButton">Search</button>
+          </div>
+      </div>
+  )
+  
 }
 
 export default Home;
